+++ src/data/trajectory_structure.hpp
@@ -7,7 +7,7 @@ namespace data_structures {
 
     struct Location {
         int order{};
-        long double timestamp{};
+        long timestamp{};
         double longitude{};
         double latitude{};
