+++ src/trajectory_data_handling/trajectory_file_manager.hpp
@@ -13,7 +13,7 @@ namespace trajectory_data_handling {
        const std::filesystem::path TDRIVE_PATH = std::filesystem::current_path().parent_path().parent_path() / "external" / "datasets" /  "t-drive";
        const std::filesystem::path GEOLIFE_PATH = std::filesystem::current_path().parent_path().parent_path() / "external" / "datasets" / "geolife";
        const char DELIMITER = ',';
-       static long double stringToTime(const std::string& timeString);
+       static long stringToTime(const std::string& timeString);
    public:
        void load_tdrive_dataset(std::vector<data_structures::Trajectory> &all_trajectories);
        void load_geolife_dataset(std::vector<data_structures::Trajectory> &all_trajectories);
